requirements.txt
# Minimal Claude CLI Integration - Free Version
# Only essential dependencies

fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
structlog>=23.2.0
pydantic>=2.0.0
anthropic>=0.7.0
